build resource based authorization

+++ Ecom.Controller/src/Controller/Ordercontroller.cs
@@ -1,9 +1,12 @@
-using Ecom.Core.src.Entity;
+using System.Security.Claims;
+using Ecom.Core.src.Enum;
 using Ecom.Core.src.parameters;
 using Ecom.Service.src.Abstraction;
 using Ecom.Service.src.DTO;
+using Ecom.Service.src.Shared;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Routing;
 
 namespace Ecom.Controller.src.Controller
 {
@@ -12,10 +15,12 @@ namespace Ecom.Controller.src.Controller
     public class Ordercontroller : ControllerBase
     {
         private readonly IOrderService _orderService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public Ordercontroller(IOrderService orderService)
+        public Ordercontroller(IOrderService orderService, IAuthorizationService authorizationService)
         {
             _orderService = orderService;
+            _authorizationService = authorizationService;
         }
 
         [Authorize(Roles = "Admin")]
@@ -26,32 +31,89 @@ namespace Ecom.Controller.src.Controller
             return Ok(result);
         }
 
-        [Authorize(Roles = "User, Admin")]
+        [Authorize]
         [HttpGet("{orderId}")]
         public async Task<ActionResult<IEnumerable<OrderReadDTO>>> GetOneOrderAsync(Guid orderId)
         {
-            var result = await _orderService.GetOneOrderAsync(orderId);
-            return Ok(result);
+            var order = await _orderService.GetOneOrderAsync(orderId) ?? throw CustomException.NotFoundException("order not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(HttpContext.User, order, "OrderOwnerPolicy");
+            if (authorizationResult.IsCompletedSuccessfully)
+            {
+                return Ok(order);
+            }
+            else
+            {
+                throw CustomException.ForbiddenException("Only User or Admin has permission to access this property");
+            }
         }
+
 
         [Authorize(Roles = "User")]
         [HttpPost()]
         public async Task<ActionResult<OrderReadDTO>> CreateOrderAsync([FromBody] OrderCreateDTO orderCreateDTO)
         {
+            if (orderCreateDTO == null)
+            {
+                throw CustomException.BadRequestException("Invalid order data");
+            }
+
+            var userClaim = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier);
+
+            if (!Guid.TryParse(userClaim?.Value, out var userId))
+            {
+                throw CustomException.BadRequestException("Invalid user ID format");
+            }
+
+            if (userId != orderCreateDTO.UserId)
+            {
+                throw CustomException.ForbiddenException("Order owner mismatch with the logged-in customer");
+            }
+
             var result = await _orderService.CreateOrderAsync(orderCreateDTO);
             return Ok(result);
         }
 
-        [Authorize(Roles = "Admin")]
+
+
+        [Authorize]
         [HttpDelete("{orderId}")]
         public async Task<ActionResult<bool>> DeleteOrderAsync(Guid orderId)
         {
-            var result = await _orderService.DeleteOrderAsync(orderId);
-            return StatusCode(204, result);
+
+            var order = await _orderService.GetOneOrderAsync(orderId) ?? throw CustomException.NotFoundException("No order is found with this id");
+
+            if (order.OrderStatus != OrderStatus.CANCELED)
+            {
+                throw CustomException.ForbiddenException("Order must be confirmed as canceled");
+            }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(HttpContext.User, order, "OrderAdminOrOwnerPolicy");
+
+            Console.WriteLine(authorizationResult.ToString());
+
+
+            if (authorizationResult.Succeeded)
+            {
+                var result = await _orderService.DeleteOrderAsync(orderId);
+                return StatusCode(204, result);
+
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+
+                return new ForbidResult();
+            }
+            else
+            {
+                return new ChallengeResult();
+
+            }
+
         }
 
 
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         [HttpPatch("{orderId}")]
         public async Task<ActionResult<OrderReadDTO>> UpdateOrderAsync(Guid orderId, [FromBody] OrderUpdateDTO orderUpdateDTO)
         {
